Header styled and updated
