+++ frontend/src/components/Header/Header.scss
@@ -1,0 +1,7 @@
+header {
+  background-color: #efefef;
+  padding: 1rem 2rem;
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 4rem;
+}
