+++ frontend/src/components/Header/Logo.js
@@ -1,4 +1,3 @@
-import react from "react-router-dom";
 
 const Logo = () => {
   return (
