+++ frontend/src/components/index.js
@@ -1,5 +1,5 @@
-export { default as Navbar } from './Header/Navbar';
-export { default as PrivateRoute } from './PrivateRoute';
-export { default as Classes } from './ClassesPage/Classes';
-export { default as Class } from './ClassPage/Class';
-export { default as Home } from './HomePage/HomePage';
+export { default as Header } from "./Header/Header";
+export { default as PrivateRoute } from "./PrivateRoute";
+export { default as Classes } from "./ClassesPage/Classes";
+export { default as Class } from "./ClassPage/Class";
+export { default as Home } from "./HomePage/HomePage";
